ai_service/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
numpy
python-dotenv
